requirements.txt
# =============================================================================
# Trading Bot — Python Dependencies
# Python 3.11+ required
# =============================================================================

# -----------------------------------------------------------------------------
# Broker — Angel One SmartAPI
# -----------------------------------------------------------------------------
smartapi-python>=1.3.4
pyotp>=2.9.0          # TOTP generation for 2FA login

# -----------------------------------------------------------------------------
# Configuration & environment
# -----------------------------------------------------------------------------
pyyaml>=6.0.1
python-dotenv>=1.0.1

# -----------------------------------------------------------------------------
# Data & technical analysis
# -----------------------------------------------------------------------------
pandas>=2.1.0
numpy>=1.26.0
pandas-ta>=0.3.14b    # Pure-Python TA library (no TA-Lib C dependency)
yfinance>=0.2.40      # Global index / fallback OHLCV data

# -----------------------------------------------------------------------------
# LLM providers
# -----------------------------------------------------------------------------
google-genai>=1.0.0              # Gemini (new SDK — replaces deprecated google-generativeai)
groq>=0.9.0                      # Groq (LLaMA)
openai>=1.40.0                   # NVIDIA NIM uses OpenAI-compatible API

# -----------------------------------------------------------------------------
# Agent framework
# -----------------------------------------------------------------------------
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langgraph>=0.1.0

# -----------------------------------------------------------------------------
# Vector memory
# -----------------------------------------------------------------------------
chromadb>=0.5.0

# -----------------------------------------------------------------------------
# Web scraping & news
# -----------------------------------------------------------------------------
requests>=2.32.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml>=5.2.0           # Fast HTML parser used by bs4 and newspaper3k
lxml-html-clean>=0.1.0  # lxml.html.clean was split into a separate package (required by newspaper3k)
feedparser>=6.0.0     # RSS/Atom feed parsing (Google News, MoneyControl)
ddgs>=0.3.0               # Free web search, no API key needed (package was renamed from duckduckgo-search)

# -----------------------------------------------------------------------------
# Search APIs
# -----------------------------------------------------------------------------
tavily-python>=0.3.0
google-search-results>=2.4.2     # SerpAPI client

# -----------------------------------------------------------------------------
# Notifications
# -----------------------------------------------------------------------------
python-telegram-bot>=21.0

# -----------------------------------------------------------------------------
# Scheduling
# -----------------------------------------------------------------------------
apscheduler>=3.10.0

# -----------------------------------------------------------------------------
# Backtesting
# -----------------------------------------------------------------------------
backtrader>=1.9.78.123

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=8.2.0
pytest-cov>=5.0.0
